refactor(protocol): migrate MCU payload serialization from Protobuf/nanopb to MsgPack
Replace nanopb-based protobuf serialization with a minimal header-only
MsgPack codec (msgpack_codec.h) and auto-generated struct encode/decode
methods (rpc_structs.h.j2). This removes the nanopb dependency entirely.

MCU changes:
- Add msgpack_codec.h: bounds-checked Encoder/Decoder with sticky _ok flag
- Rewrite rpc_structs.h template: strnlen (defense-in-depth), array count
  validation on decode, [[nodiscard]] on zero-copy read_bin_view()
- Simplify Bridge.h templates: remove _dispatchWithBytes, is_two_arg_invocable
- Rewrite sendPbFrame/sendPbCommand/emitStatus to use msgpack::Encoder
- FileSystem: _onWrite(msg,data) -> _onWrite(msg), data in msg via span
- Process: widen PID type from int16_t to int32_t (no truncation risk)
- SPI transfer: copy to mutable buffer instead of const_cast UB
- Remove pinRead misleading static_cast<uint32_t>
- Rename pb_copy.h -> string_copy.h

Python changes:
- Replace protobuf Packet classes with msgspec.Struct (array_like=True)
- Remove protobuf/grpcio from runtime.toml and runtime.txt
- Update services to use new encode/decode patterns

Code generation:
- Remove duplicate class/function definitions in generate.py
- Template emits strnlen instead of strlen for str_fixed fields
- Template validates array element count on decode

Build & CI:
- Remove nanopb sources/includes from all build scripts
- Remove install_nanopb() from install.sh
- Update all 8 documentation files

Verified: 8 C++ suites (115 tests, 0 failures), 811 Python tests passed,
8/8 e2e client scripts passed.

+++ mcubridge-library-arduino/src/services/Process.h
@@ -21,7 +21,7 @@ class ProcessClass : public BridgeObserver {
   friend class bridge::test::ProcessTestAccessor;
 #endif
  public:
-  using ProcessRunAsyncHandler = etl::delegate<void(int16_t)>;
+  using ProcessRunAsyncHandler = etl::delegate<void(int32_t)>;
   using ProcessPollHandler = etl::delegate<void(rpc::StatusCode, uint8_t, etl::span<const uint8_t>, etl::span<const uint8_t>)>;
   using ProcessKillHandler = etl::delegate<void(rpc::StatusCode)>;
 
@@ -32,8 +32,8 @@ class ProcessClass : public BridgeObserver {
   [[maybe_unused]] void notification(MsgBridgeLost) override { reset(); }
 
   [[maybe_unused]] void runAsync(etl::string_view command, etl::span<const etl::string_view> args, ProcessRunAsyncHandler handler);
-  [[maybe_unused]] void poll(int16_t pid, ProcessPollHandler handler);
-  [[maybe_unused]] void kill(int16_t pid, ProcessKillHandler handler = {});
+  [[maybe_unused]] void poll(int32_t pid, ProcessPollHandler handler);
+  [[maybe_unused]] void kill(int32_t pid, ProcessKillHandler handler = {});
   bool _kill(uint32_t pid);
   void reset();
 
@@ -46,7 +46,7 @@ class ProcessClass : public BridgeObserver {
   };
 
   struct PendingPoll {
-    int16_t pid;
+    int32_t pid;
     ProcessPollHandler handler;
   };
